re-organize python code (#6)
