+++ Cargo.toml
@@ -14,7 +14,7 @@ build = "build.rs"
 [dependencies]
 datafusion = { git = "https://github.com/apache/arrow-datafusion", rev = "5238e8c97f998b4d2cb9fab85fb182f325a1a7fb", features = ["pyarrow", "avro"] }
 datafusion-proto = { git = "https://github.com/apache/arrow-datafusion", rev = "5238e8c97f998b4d2cb9fab85fb182f325a1a7fb" }
-datafusion-python = { path = "../../apache/arrow-datafusion-python" }
+datafusion-python = { git = "https://github.com/apache/arrow-datafusion-python", rev = "0a6743c2a7f27015890c5546efda2d4e85d45a97" }
 futures = "0.3"
 prost = "0.11"
 prost-types = "0.11"
